refactor: nanocl setup command
