+++ bin/nanocl/src/commands/state.rs
@@ -1,5 +1,6 @@
 use std::fs;
 
+use ntex::rt;
 use clap::{Command, Arg};
 use dialoguer::Confirm;
 use dialoguer::theme::ColorfulTheme;
@@ -15,13 +16,10 @@ use nanocld_client::stubs::cargo_config::{
 use nanocld_client::stubs::state::{
   StateConfig, StateDeployment, StateCargo, StateStream,
 };
-use ntex::rt::{self, JoinHandle};
 
 use crate::utils;
 use crate::error::CliError;
 use crate::models::{StateArgs, StateCommands, StateOpts, StateBuildArgs};
-use crate::utils::print::print_yml;
-use crate::utils::url::parse_url;
 
 use super::cargo_image::exec_cargo_image_create;
 
@@ -154,7 +152,7 @@ async fn attach_to_cargo(
   client: &NanocldClient,
   cargo: CargoConfigPartial,
   namespace: &str,
-) -> Result<Vec<JoinHandle<()>>, CliError> {
+) -> Result<Vec<rt::JoinHandle<()>>, CliError> {
   let cargo = match client
     .inspect_cargo(&cargo.name, Some(namespace.to_owned()))
     .await
@@ -354,7 +352,7 @@ fn inject_build_args(
 }
 
 async fn exec_state_apply(opts: &StateOpts) -> Result<(), CliError> {
-  let (meta, yaml) = match parse_url(&opts.file_path) {
+  let (meta, yaml) = match utils::url::parse_url(&opts.file_path) {
     Ok(url) => get_from_url(url).await?,
     Err(_) => get_from_file(&opts.file_path).await?,
   };
@@ -385,7 +383,7 @@ async fn exec_state_apply(opts: &StateOpts) -> Result<(), CliError> {
     _ => yaml,
   };
   let data = serde_json::to_value(&yaml)?;
-  let _ = print_yml(&yaml);
+  let _ = utils::print::print_yml(&yaml);
   if !opts.skip_confirm {
     let result = Confirm::with_theme(&ColorfulTheme::default())
       .with_prompt("Are you sure to apply this new state ?")
@@ -417,7 +415,7 @@ async fn exec_state_apply(opts: &StateOpts) -> Result<(), CliError> {
 }
 
 async fn exec_state_revert(opts: &StateOpts) -> Result<(), CliError> {
-  let (meta, yaml) = match parse_url(&opts.file_path) {
+  let (meta, yaml) = match utils::url::parse_url(&opts.file_path) {
     Ok(url) => get_from_url(url).await?,
     Err(_) => get_from_file(&opts.file_path).await?,
   };
@@ -426,7 +424,7 @@ async fn exec_state_revert(opts: &StateOpts) -> Result<(), CliError> {
 
   let yaml = inject_build_args(yaml.clone(), opts.args.clone())?;
   let data = serde_json::to_value(&yaml)?;
-  let _ = print_yml(&yaml);
+  let _ = utils::print::print_yml(&yaml);
   if !opts.skip_confirm {
     let result = Confirm::with_theme(&ColorfulTheme::default())
       .with_prompt("Are you sure to revert this state ?")
