+++ bin/nanocl/src/utils/mod.rs
@@ -3,3 +3,4 @@ pub mod print;
 pub mod math;
 pub mod network;
 pub mod url;
+pub mod docker;
